added a button to take focus
i added a button that is hidden but is able to take focus. I did this to be able to clear focus from the slider element after the user stops dragging it. I tried to find a better solution but i came up with this myself because i couldnt find anything